pyproject.toml
[project]
name = "first-gen-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
   # "google-generativeai (>=0.8.5,<0.9.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "google (>=3.0.0,<4.0.0)",
    "openai (>=1.97.2,<2.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-google-genai==2.1.3",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "langchain-google-vertexai (>=2.0.27,<3.0.0)",
    "langgraph (>=0.6.2,<0.7.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "duckduckgo-search (>=8.1.1,<9.0.0)",
    "ddgs (>=9.4.3,<10.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "pypdf (>=5.9.0,<6.0.0)",
    "chroma (>=0.2.0,<0.3.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "sqlite4 (>=0.1.1,<0.2.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "tabulate (>=0.9.0,<0.10.0)"
]

[tool.poetry]
packages = [{include = "first_gen_ai", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
